+++ 1.21.11/src/main/java/com/bleudev/nine_lifes/mixin/CampfireBlockEntityMixin.java
@@ -1,5 +1,6 @@
 package com.bleudev.nine_lifes.mixin;
 
+import com.bleudev.nine_lifes.custom.NineLifesDamageSources;
 import com.bleudev.nine_lifes.custom.NineLifesItemTags;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.NonNullList;
@@ -11,7 +12,6 @@ import net.minecraft.world.item.crafting.SingleRecipeInput;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.CampfireBlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -19,6 +19,8 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import static com.bleudev.nine_lifes.util.InjectsKt.explode;
 
 @Mixin(CampfireBlockEntity.class)
 public abstract class CampfireBlockEntityMixin {
@@ -32,7 +34,7 @@ public abstract class CampfireBlockEntityMixin {
     private int[] cookingTime;
 
     @Inject(method = "cookTick", at = @At("TAIL"))
-    private static void explode(
+    private static void explodeOnEnd(
         ServerLevel level, BlockPos pos, BlockState state, CampfireBlockEntity campfire,
         RecipeManager.CachedCheck<@NotNull SingleRecipeInput, @NotNull CampfireCookingRecipe> recipe,
         CallbackInfo ci) {
@@ -42,9 +44,8 @@ public abstract class CampfireBlockEntityMixin {
             ItemStack stack = self.getItems().get(slot);
             if (stack.isEmpty() || !stack.is(NineLifesItemTags.CAUSE_CAMPFIRE_EXPLODE)) continue;
             if (self.cookingProgress[slot] == self.cookingTime[slot] - 1) {
-                Vec3 center = pos.getCenter();
                 level.removeBlock(pos, false);
-                level.explode(null, center.x, center.y, center.z, 4f, true, Level.ExplosionInteraction.BLOCK);
+                explode(level, pos.getCenter(), 4f, NineLifesDamageSources::charged, Level.ExplosionInteraction.BLOCK, null);
                 return;
             }
         }