+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Fixes
 - All nine lifes damage types do not knockback (bleudev)
+- Use charged amethyst damage source in amethyst explosions
 
 ## Optimisations
 - Clean some mixins